Refactored project structure: added dao, entity, and service directories under model. Reorganized model classes into respective directories.
Refactored classes in the entity directory
