+++ CoworkingSpaceReservation/src/main/java/org/example/controller/AdminController.java
@@ -1,6 +1,9 @@
 package org.example.controller;
 
-import org.example.model.*;
+import org.example.model.entity.Reservation;
+import org.example.model.entity.User;
+import org.example.model.entity.Workspace;
+import org.example.model.service.AdminManager;
 import org.example.view.AdminView;
 
 import java.util.List;
