+++ CoworkingSpaceReservation/src/main/java/org/example/controller/ReservationController.java
@@ -1,6 +1,10 @@
 package org.example.controller;
 
-import org.example.model.*;
+import org.example.model.entity.Customer;
+import org.example.model.entity.Reservation;
+import org.example.model.entity.User;
+import org.example.model.entity.Workspace;
+import org.example.model.service.ReservationManager;
 
 import java.time.LocalDateTime;
 import java.util.List;
