+++ CoworkingSpaceReservation/src/main/java/org/example/controller/navigation/command/customer/CustomerCommand.java
@@ -2,7 +2,7 @@ package org.example.controller.navigation.command.customer;
 
 import lombok.Getter;
 import org.example.controller.navigation.command.MenuCommand;
-import org.example.model.Customer;
+import org.example.model.entity.Customer;
 @Getter
 public abstract class CustomerCommand implements MenuCommand {
     private final Customer customer;
