+++ CoworkingSpaceReservation/src/main/java/org/example/controller/navigation/command/customer/MakeReservationCommand.java
@@ -4,7 +4,7 @@ import org.example.controller.GeneralController;
 import org.example.controller.ValueValidator;
 import org.example.exceptions.IdNotFoundException;
 import org.example.exceptions.ReservationAlreadyExistException;
-import org.example.model.Customer;
+import org.example.model.entity.Customer;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
