+++ CoworkingSpaceReservation/src/main/java/org/example/controller/navigation/command/customer/ShowAvailableSpacesCommand.java
@@ -2,7 +2,7 @@ package org.example.controller.navigation.command.customer;
 
 import org.example.controller.GeneralController;
 import org.example.controller.ValueValidator;
-import org.example.model.Customer;
+import org.example.model.entity.Customer;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
