--- CoworkingSpaceReservation/src/main/java/org/example/model/User.java
@@ -1,18 +1,0 @@
-package org.example.model;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.*;
-import lombok.experimental.SuperBuilder;
-
-
-@Data
-@SuperBuilder
-public abstract class User {
-    @JsonProperty("login")
-    private String login;
-    @JsonCreator
-    User(String login) {
-        this.login = login;
-    }
-}
