+++ CoworkingSpaceReservation/src/main/java/org/example/model/entity/Customer.java
@@ -1,0 +1,12 @@
+package org.example.model.entity;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+@EqualsAndHashCode(callSuper = true)
+@Data
+@SuperBuilder
+@NoArgsConstructor
+public class Customer extends User {
+}
