+++ CoworkingSpaceReservation/src/main/java/org/example/model/entity/Reservation.java
@@ -1,4 +1,4 @@
-package org.example.model;
+package org.example.model.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
