+++ CoworkingSpaceReservation/src/main/java/org/example/model/entity/User.java
@@ -1,0 +1,12 @@
+package org.example.model.entity;
+
+import lombok.*;
+import lombok.experimental.SuperBuilder;
+
+@Data
+@SuperBuilder
+@NoArgsConstructor
+public abstract class User {
+    private long id;
+    private String login;
+}
