--- CoworkingSpaceReservation/src/main/java/org/example/model/factory/UserFactory.java
@@ -1,7 +1,0 @@
-package org.example.model.factory;
-
-import org.example.model.User;
-
-interface UserFactory {
-    User createUser();
-}
