+++ CoworkingSpaceReservation/src/main/java/org/example/model/service/CustomerManager.java
@@ -1,4 +1,7 @@
-package org.example.model;
+package org.example.model.service;
+
+import org.example.model.entity.Customer;
+import org.example.model.entity.User;
 
 public class CustomerManager {
