+++ CoworkingSpaceReservation/src/main/java/org/example/model/service/LoginManager.java
@@ -1,6 +1,9 @@
-package org.example.model;
+package org.example.model.service;
 
 
+import org.example.model.entity.Admin;
+import org.example.model.entity.Customer;
+import org.example.model.entity.User;
 import org.example.model.storage.ApplicationStateManager;
 
 import java.util.Map;
