+++ CoworkingSpaceReservation/src/main/java/org/example/model/service/factory/AdminFactory.java
@@ -1,8 +1,8 @@
-package org.example.model.factory;
+package org.example.model.service.factory;
 
 
-import org.example.model.AdminManager;
-import org.example.model.User;
+import org.example.model.service.AdminManager;
+import org.example.model.entity.User;
 
 class AdminFactory implements UserFactory {
     @Override
