+++ CoworkingSpaceReservation/src/main/java/org/example/model/service/factory/CustomerFactory.java
@@ -1,7 +1,7 @@
-package org.example.model.factory;
+package org.example.model.service.factory;
 
-import org.example.model.CustomerManager;
-import org.example.model.User;
+import org.example.model.service.CustomerManager;
+import org.example.model.entity.User;
 
 class CustomerFactory implements UserFactory {
     @Override
