+++ CoworkingSpaceReservation/src/main/java/org/example/model/service/factory/UserFactory.java
@@ -1,0 +1,7 @@
+package org.example.model.service.factory;
+
+import org.example.model.entity.User;
+
+interface UserFactory {
+    User createUser();
+}
