+++ CoworkingSpaceReservation/src/test/java/org/example/model/WorkspaceManagerTest.java
@@ -3,6 +3,11 @@ package org.example.model;
 import lombok.SneakyThrows;
 import org.example.exceptions.IdNotFoundException;
 import org.example.exceptions.PlaceAlreadyExistException;
+import org.example.model.entity.Reservation;
+import org.example.model.entity.SpaceType;
+import org.example.model.entity.Workspace;
+import org.example.model.service.ReservationManager;
+import org.example.model.service.WorkspaceManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
